Huge update implementing many new features.
Huge updates are bound to come as long the game isn't on a first version.

Anyways some of the new stuff:
- Skills can now be chosen in the SkillScreen.
- Language files handle languages
- Languages can be changed in the settings menu.
- Buffs and debuffs such as dots (e.g. bleeding) implemented
- New skills
- New skill graphics
- Attributes more or less implemented (skills still need to implement
  attribute scaling)
- Mana can not grow past the total you have.
- Players can be stunned now (a.k.a. unable to do sth)
- Buff system is now implemented.

This is a huge success in the development but the game has a long way to
go. Most work missing are skills and graphics unfortunately.

** Milestone note: We're almost at the first milestone having a
functioning single player game. Implementing two players on one device
should be no problem at all. Guess we'll have to speak words to continue
until then I'll try to get as far as possible.

Known bugs:

- When applying a skill to your players skill bar in the SkillScreen the
graphic is drawn before the new event coords are applied making it jump.
- Some skills aren't working as intended. This needs some testing and fine
tweeking.