[FEATURE] Implemented the Attribute Screen.
Now the Attribute Screen is implemented and with it the attributes can be
set by the player. The role is also temporally selectable in that menu
where the secondary role will be choosable.
Also fixed the stun problem that stunning skills stun infinitely.
Ascension now works too. When you beat all 10 enemies you'll be ascended
which means the enemy level increases by 1.
The enemies do not have skills according to level 3 or higher which means
they sometimes pause. This will be changed when more skills are available.